# extend Block validation timing logs to measure just the CLVM and conditions