--- a/tests/core/full_node/stores/test_full_node_store.py
+++ b/tests/core/full_node/stores/test_full_node_store.py
@@ -122,7 +122,9 @@ async def test_basic_store(
     # Add/get unfinished block
     for height, unf_block in enumerate(unfinished_blocks):
         assert store.get_unfinished_block(unf_block.partial_hash) is None
-        store.add_unfinished_block(uint32(height), unf_block, PreValidationResult(None, uint64(123532), None, False))
+        store.add_unfinished_block(
+            uint32(height), unf_block, PreValidationResult(None, uint64(123532), None, False, uint32(0))
+        )
         assert store.get_unfinished_block(unf_block.partial_hash) == unf_block
         store.remove_unfinished_block(unf_block.partial_hash)
         assert store.get_unfinished_block(unf_block.partial_hash) is None